+++ src/main/java/duke/Main.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.data.Duke;
+
 public class Main {
 
     public static void main(String[] args) {
