+++ src/main/java/duke/commands/DeleteCommand.java
@@ -1,3 +1,10 @@
+package duke.commands;
+
+import duke.data.TaskList;
+import duke.storage.Storage;
+import duke.ui.Ui;
+
+
 public class DeleteCommand extends Command{
 
     @Override
