+++ src/main/java/duke/commands/HelpCommand.java
@@ -1,3 +1,9 @@
+package duke.commands;
+
+import duke.data.TaskList;
+import duke.storage.Storage;
+import duke.ui.Ui;
+
 public class HelpCommand extends Command{
 
     @Override
