+++ src/main/java/duke/data/Duke.java
@@ -1,4 +1,13 @@
+package duke.data;
+
 import java.util.Scanner;
+
+import duke.commands.Command;
+import duke.data.TaskList;
+import duke.data.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.ui.Ui;
 
 public class Duke {
     private TaskList taskList;
@@ -6,10 +15,10 @@ public class Duke {
     private Storage storage;
 
     public Duke() {
-        this.taskList = new TaskList();
+        ;
         this.ui = new Ui(new Scanner(System.in));
         this.storage = new Storage();
-        storage.load();
+        this.taskList = this.storage.load();
     }
 
     public void run() {
@@ -24,6 +33,8 @@ public class Duke {
                 terminate = c.isEnd();
             } catch (DukeException de) {
                 System.out.println(de.getMessage());
+            } catch (IndexOutOfBoundsException ioobe) {
+                System.out.println("Index input is greater than todolist length");
             }
         } while (!terminate);
     }
