+++ src/main/java/duke/task/Deadline.java
@@ -1,10 +1,13 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Event extends Task{
+public class Deadline extends Task{
+
     protected LocalDateTime dateAndTime;
 
-    public Event(String name, String dateAndTime) {
+    public Deadline(String name, String dateAndTime) {
         super(name);
 
         String[] dateAndTimeSplit = dateAndTime.split(" ");
@@ -19,8 +22,7 @@ public class Event extends Task{
         );
     }
 
-
-    public Event(String name, String dateAndTime, boolean isDone) {
+    public Deadline(String name, String dateAndTime, boolean isDone) {
         super(name, isDone);
         String[] dateAndTimeSplit = dateAndTime.split(" ");
         String[] dateSplit = dateAndTimeSplit[0].split("[,./-]");
@@ -34,21 +36,22 @@ public class Event extends Task{
         );
     }
 
-    public String tag() {
-        return "E";
-    }
 
     public String getDateString() {
         return this.dateAndTime.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
     }
 
+    public String tag() {
+        return "D";
+    }
+
     @Override
     public String toString() {
-        return String.format("[%s]%s (at: %s)",tag(), super.toString(), getDateString());
+        return String.format("[%s]%s (by: %s)", tag(), super.toString(), getDateString());
     }
 
     @Override
     public String savedString() {
-        return String.format("%s,%s,%s", tag(), super.savedString(), this.getDateString());
+        return String.format("%s,%s,%s", tag(), super.savedString(), getDateString());
     }
 }
