+++ src/main/java/duke/task/Event.java
@@ -1,11 +1,12 @@
+package duke.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
-public class Deadline extends Task{
-
+public class Event extends Task{
     protected LocalDateTime dateAndTime;
 
-    public Deadline(String name, String dateAndTime) {
+    public Event(String name, String dateAndTime) {
         super(name);
 
         String[] dateAndTimeSplit = dateAndTime.split(" ");
@@ -20,7 +21,8 @@ public class Deadline extends Task{
         );
     }
 
-    public Deadline(String name, String dateAndTime, boolean isDone) {
+
+    public Event(String name, String dateAndTime, boolean isDone) {
         super(name, isDone);
         String[] dateAndTimeSplit = dateAndTime.split(" ");
         String[] dateSplit = dateAndTimeSplit[0].split("[,./-]");
@@ -34,22 +36,21 @@ public class Deadline extends Task{
         );
     }
 
+    public String tag() {
+        return "E";
+    }
 
     public String getDateString() {
         return this.dateAndTime.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm"));
     }
 
-    public String tag() {
-        return "D";
-    }
-
     @Override
     public String toString() {
-        return String.format("[%s]%s (by: %s)", tag(), super.toString(), getDateString());
+        return String.format("[%s]%s (at: %s)",tag(), super.toString(), getDateString());
     }
 
     @Override
     public String savedString() {
-        return String.format("%s,%s,%s", tag(), super.savedString(), getDateString());
+        return String.format("%s,%s,%s", tag(), super.savedString(), this.getDateString());
     }
 }
